Updated relations, optimized import operation.
