+++ src/api/microorganism/content-types/microorganism/schema.json
@@ -18,10 +18,10 @@
     "iri": {
       "type": "string"
     },
-    "labtests": {
+    "isolates": {
       "type": "relation",
       "relation": "oneToMany",
-      "target": "api::labtest.labtest",
+      "target": "api::isolate.isolate",
       "mappedBy": "microorganism"
     }
   }
