--- src/api/objective/controllers/objective.ts
@@ -1,7 +1,0 @@
-/**
- * objective controller
- */
-
-import { factories } from '@strapi/strapi'
-
-export default factories.createCoreController('api::objective.objective');
