--- src/api/objective/routes/objective.ts
@@ -1,7 +1,0 @@
-/**
- * objective router
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreRouter('api::objective.objective');
