--- src/api/objective/services/objective.ts
@@ -1,7 +1,0 @@
-/**
- * objective service
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreService('api::objective.objective');
