+++ src/api/sampling-objective/content-types/sampling-objective/schema.json
@@ -1,10 +1,11 @@
 {
   "kind": "collectionType",
-  "collectionName": "matrices",
+  "collectionName": "sampling_objectives",
   "info": {
-    "singularName": "matrix",
-    "pluralName": "matrices",
-    "displayName": "Matrix"
+    "singularName": "sampling-objective",
+    "pluralName": "sampling-objectives",
+    "displayName": "Sampling Objective",
+    "description": ""
   },
   "options": {
     "draftAndPublish": false
@@ -21,7 +22,7 @@
       "type": "relation",
       "relation": "oneToMany",
       "target": "api::isolate.isolate",
-      "mappedBy": "matrix"
+      "mappedBy": "sampling_objective"
     }
   }
 }
