+++ src/api/sampling-objective/controllers/sampling-objective.ts
@@ -1,0 +1,7 @@
+/**
+ * sampling-objective controller
+ */
+
+import { factories } from '@strapi/strapi'
+
+export default factories.createCoreController('api::sampling-objective.sampling-objective');
