+++ src/api/sampling-objective/routes/sampling-objective.ts
@@ -1,0 +1,7 @@
+/**
+ * sampling-objective router
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::sampling-objective.sampling-objective');
