+++ src/api/sampling-objective/services/sampling-objective.ts
@@ -1,0 +1,7 @@
+/**
+ * sampling-objective service
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreService('api::sampling-objective.sampling-objective');
