--- src/api/welcome/content-types/welcome/schema.json
@@ -1,35 +1,0 @@
-{
-  "kind": "singleType",
-  "collectionName": "welcomes",
-  "info": {
-    "singularName": "welcome",
-    "pluralName": "welcomes",
-    "displayName": "welcome"
-  },
-  "options": {
-    "draftAndPublish": true
-  },
-  "pluginOptions": {
-    "i18n": {
-      "localized": true
-    }
-  },
-  "attributes": {
-    "subheading": {
-      "pluginOptions": {
-        "i18n": {
-          "localized": true
-        }
-      },
-      "type": "text"
-    },
-    "content": {
-      "pluginOptions": {
-        "i18n": {
-          "localized": true
-        }
-      },
-      "type": "richtext"
-    }
-  }
-}
