--- src/api/welcome/controllers/welcome.ts
@@ -1,7 +1,0 @@
-/**
- * welcome controller
- */
-
-import { factories } from '@strapi/strapi'
-
-export default factories.createCoreController('api::welcome.welcome');
