--- src/api/welcome/routes/welcome.ts
@@ -1,7 +1,0 @@
-/**
- * welcome router
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreRouter('api::welcome.welcome');
