--- src/api/welcome/services/welcome.ts
@@ -1,7 +1,0 @@
-/**
- * welcome service
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreService('api::welcome.welcome');
