--- src/api/zoonotify/controllers/zoonotify.ts
@@ -1,7 +1,0 @@
-/**
- * zoonotify controller
- */
-
-import { factories } from '@strapi/strapi'
-
-export default factories.createCoreController('api::zoonotify.zoonotify');
