--- src/api/zoonotify/services/zoonotify.ts
@@ -1,7 +1,0 @@
-/**
- * zoonotify service
- */
-
-import { factories } from '@strapi/strapi';
-
-export default factories.createCoreService('api::zoonotify.zoonotify');
